+++ test/dummy/app/models/wearer/palette_snapshot.rb
@@ -1,0 +1,3 @@
+class Wearer::PaletteSnapshot < ApplicationRecord
+  include Pandatone::Dresser::Snapshot
+end